- ajout de la synchronisation des 9PG.

+++ module/ebdj-moteur/src/main/java/fr/qp1c/ebdj/moteur/dao/DBConnecteurSynchroDao.java
@@ -2,7 +2,7 @@ package fr.qp1c.ebdj.moteur.dao;
 
 public interface DBConnecteurSynchroDao {
 
-	public Integer recupererIndexParCle(String cle);
+	public Long recupererIndexParCle(String cle);
 
-	public void modifierIndexParCle(String cle, Integer nouvelIndex);
+	public void modifierIndexParCle(String cle, Long nouvelIndex);
 }
